K_mean Clustering
The coded has two functions:
The first function (initialization_random): randomly according to the number k; Specifies the centers. For example, if k is 3, it randomly selects 3 centers from the data.
Second function (centruids_update): For each data, it checks which data center is closer to the first function. If it was close to one, then that data belongs to that center.
The same goes for all data. After categorizing the data; For each category, an average is taken between the data and it is considered as the new center of that category. This process continues until the algorithm converges and gives the final result.
These categories are also drawn in the last piece of code.